Resolve JD-1674 "3"

See merge request ATL-TEAM/jira-plugins-mrimsender!50

+++ src/main/java/ru/mail/jira/plugins/myteam/bot/rulesengine/states/base/CancelableState.java
@@ -1,8 +1,8 @@
 /* (C)2021 */
 package ru.mail.jira.plugins.myteam.bot.rulesengine.states.base;
 
-import com.atlassian.jira.user.ApplicationUser;
 import java.io.IOException;
+import org.jetbrains.annotations.Nullable;
 import ru.mail.jira.plugins.myteam.bot.events.ButtonClickEvent;
 import ru.mail.jira.plugins.myteam.bot.events.MyteamEvent;
 import ru.mail.jira.plugins.myteam.commons.exceptions.MyteamServerErrorException;
@@ -10,6 +10,7 @@ import ru.mail.jira.plugins.myteam.service.UserChatService;
 
 public interface CancelableState {
 
+  @Nullable
   default UserChatService getUserChatService() {
     return null;
   }
@@ -18,21 +19,23 @@ public interface CancelableState {
 
   default void cancel(MyteamEvent event) {
     UserChatService userChatService = getUserChatService();
-    userChatService.deleteState(event.getChatId());
+    if (userChatService != null) userChatService.deleteState(event.getChatId());
 
     if (event instanceof ButtonClickEvent) {
       try {
-        ApplicationUser user = userChatService.getJiraUserFromUserChatId(event.getChatId());
-        userChatService.answerCallbackQuery(((ButtonClickEvent) event).getQueryId());
-        userChatService.editMessageText(
-            event.getChatId(),
-            ((ButtonClickEvent) event).getMsgId(),
-            userChatService.getRawText(
-                "ru.mail.jira.plugins.myteam.myteamEventsListener.actionCanceled"),
-            null);
+        if (userChatService != null) {
+          userChatService.answerCallbackQuery(((ButtonClickEvent) event).getQueryId());
+          userChatService.editMessageText(
+              event.getChatId(),
+              ((ButtonClickEvent) event).getMsgId(),
+              userChatService.getRawText(
+                  "ru.mail.jira.plugins.myteam.myteamEventsListener.actionCanceled"),
+              null);
+        }
       } catch (MyteamServerErrorException | IOException e) {
         onError(e);
       }
     }
-  };
+  }
+  ;
 }
